+++ autoprotocol/harness.py
@@ -3,23 +3,20 @@ import json
 import io
 from .protocol import Protocol
 from .unit import Unit, UnitError
-from .container import WellGroup, SEAL_TYPES, COVER_TYPES  # NOQA
+from .container import WellGroup
 from . import UserError
 import argparse
 import sys
 
-if sys.version_info[0] >= 3:
-    string_type = str
-else:
-    string_type = basestring  # pylint: disable=undefined-variable
+if sys.version_info.major == 3:
+    basestring = str  # pylint: disable=invalid-name
 
-"""
+'''
     :copyright: 2017 by The Autoprotocol Development Team, see AUTHORS
         for more details.
     :license: BSD, see LICENSE for more details
 
-"""
-
+'''
 
 _DYE_TEST_RS = {
     "dye4000": "rs18qmhr7t9jwq",
@@ -28,7 +25,7 @@ _DYE_TEST_RS = {
 
 
 def param_default(typeDesc):
-    if isinstance(typeDesc, string_type):
+    if isinstance(typeDesc, basestring):
         typeDesc = {'type': typeDesc}
 
     type = typeDesc['type']
@@ -79,7 +76,7 @@ def convert_param(protocol, val, typeDesc):
         Description of input type.
 
     """
-    if isinstance(typeDesc, string_type):
+    if isinstance(typeDesc, basestring):
         typeDesc = {'type': typeDesc}
     if val is None:
         val = param_default(typeDesc)
@@ -98,21 +95,21 @@ def convert_param(protocol, val, typeDesc):
                                "reference to an aliquot" % (val, label))
     elif type == 'aliquot+':
         try:
-            return WellGroup([convert_param(protocol, a, 'aliquot')
-                             for a in val])
+            return WellGroup(
+                [convert_param(protocol, a, 'aliquot') for a in val])
         except:
             label = typeDesc.get('label') or "[unknown]"
-            raise RuntimeError("The value supplied to input '%s' "
-                               "(type aliquot+) is improperly formatted."
-                               "" % label)
+            raise RuntimeError(
+                "The value supplied to input '%s' (type aliquot+) is "
+                "improperly formatted." % label)
     elif type == 'aliquot++':
         try:
             return [convert_param(protocol, aqs, 'aliquot+') for aqs in val]
         except:
             label = typeDesc.get('label') or "[unknown]"
-            raise RuntimeError("The value supplied to input '%s' "
-                               "(type aliquot++) is improperly formatted."
-                               "" % label)
+            raise RuntimeError(
+                "The value supplied to input '%s' (type aliquot++) is "
+                "improperly formatted." % label)
     elif type == 'container':
 
         try:
@@ -126,24 +123,24 @@ def convert_param(protocol, val, typeDesc):
             return [convert_param(protocol, cont, 'container') for cont in val]
         except:
             label = typeDesc.get('label') or "[unknown]"
-            raise RuntimeError("The value supplied to input '%s' "
-                               "(type container+) is improperly formatted."
-                               "" % label)
+            raise RuntimeError(
+                "The value supplied to input '%s' (type container+) is "
+                "improperly formatted." % label)
     elif type in ['volume', 'time', 'length', 'frequency']:
         try:
-            return Unit.fromstring(val)
+            return Unit(val)
         except UnitError as e:
             raise RuntimeError("The value supplied (%s) as a unit of '%s' is "
                                "improperly formatted. Units of %s must be in "
-                               "the form: 'number:unit'"
-                               "" % (e.value, type, type))
+                               "the form: 'number:unit'" % (
+                               e.value, type, type))
     elif type == 'temperature':
         try:
-            if val in ['ambient', 'warm_30', 'warm_37',
-                       'cold_4', 'cold_20', 'cold_80']:
+            if val in ['ambient', 'warm_30', 'warm_37', 'cold_4', 'cold_20',
+                       'cold_80']:
                 return val
             else:
-                return Unit.fromstring(val)
+                return Unit(val)
         except UnitError as e:
             raise RuntimeError("Invalid temperature value for %s: temperature "
                                "input types must be either storage conditions "
@@ -160,17 +157,17 @@ def convert_param(protocol, val, typeDesc):
             return int(val)
         except ValueError:
             label = typeDesc.get('label') or "[unknown]"
-            raise RuntimeError("The value supplied to input '%s' "
-                               "(type integer) is improperly formatted."
-                               "" % label)
+            raise RuntimeError(
+                "The value supplied to input '%s' (type integer) is improperly "
+                "formatted." % label)
     elif type == 'decimal':
         try:
             return float(val)
         except ValueError:
             label = typeDesc.get('label') or "[unknown]"
-            raise RuntimeError("The value supplied to input '%s' "
-                               "(type decimal) is improperly "
-                               "formatted." % label)
+            raise RuntimeError(
+                "The value supplied to input '%s' (type decimal) is improperly "
+                "formatted." % label)
     elif type == 'group':
         try:
             return {
@@ -179,28 +176,29 @@ def convert_param(protocol, val, typeDesc):
             }
         except KeyError as e:
             label = typeDesc.get('label') or "[unknown]"
-            raise RuntimeError("The value supplied to input '%s' "
-                               "(type group) is missing a(n) %s field."
-                               "" % (label, e))
+            raise RuntimeError(
+                "The value supplied to input '%s' (type group) is missing "
+                "a(n) %s field." % (label, e))
         except AttributeError:
             label = typeDesc.get('label') or "[unknown]"
-            raise RuntimeError("The value supplied to input '%s' "
-                               "(type group) is improperly formatted." % label)
+            raise RuntimeError(
+                "The value supplied to input '%s' (type group) is improperly "
+                "formatted." % label)
     elif type == 'group+':
         try:
             return [{
-                    k: convert_param(protocol, x.get(k), typeDesc['inputs'][k])
-                    for k in typeDesc['inputs']
-                    } for x in val]
+                k: convert_param(protocol, x.get(k), typeDesc['inputs'][k])
+                for k in typeDesc['inputs']
+            } for x in val]
         except (TypeError, AttributeError):
-            raise RuntimeError("The value supplied to input '%s' "
-                               "(type group+) must be in the form of a "
-                               "list of dictionaries" % typeDesc['label'])
+            raise RuntimeError(
+                "The value supplied to input '%s' (type group+) must be in "
+                "the form of a list of dictionaries" % typeDesc['label'])
         except KeyError as e:
             label = typeDesc.get('label') or "[unknown]"
-            raise RuntimeError("The value supplied to input '%s' "
-                               "(type group+) is missing a(n) %s field."
-                               "" % (label, e))
+            raise RuntimeError(
+                "The value supplied to input '%s' (type group+) is missing "
+                "a(n) %s field." % (label, e))
     elif type == 'group-choice':
         try:
             return {
@@ -210,21 +208,23 @@ def convert_param(protocol, val, typeDesc):
                         protocol,
                         val['inputs'].get(opt['value']),
                         {'type': 'group', 'inputs': opt['inputs']})
-                    for opt in typeDesc['options'] if opt['value'] == val['value']
+                    for opt in typeDesc['options'] if
+                opt['value'] == val['value']
                 }
             }
         except (KeyError, AttributeError) as e:
             label = typeDesc.get('label') or "[unknown]"
             if e in ["value", "inputs"]:
-                raise RuntimeError("The value supplied to input '%s' "
-                                   "(type group-choice) is missing a(n) %s "
-                                   "field." % (label, e))
+                raise RuntimeError(
+                    "The value supplied to input '%s' (type group-choice) "
+                    "is missing a(n) %s field." % (label, e))
     elif type == 'thermocycle':
         try:
             return [
                 {
                     'cycles': g['cycles'],
-                    'steps': [convert_param(protocol, s, 'thermocycle_step') for s in g['steps']]
+                    'steps': [convert_param(protocol, s, 'thermocycle_step') for
+                              s in g['steps']]
                 }
                 for g in val
             ]
@@ -233,25 +233,25 @@ def convert_param(protocol, val, typeDesc):
 
     elif type == 'thermocycle_step':
         try:
-            output = {'duration': Unit.fromstring(val['duration'])}
+            output = {'duration': Unit(val['duration'])}
         except UnitError as e:
-            raise RuntimeError("Invalid duration value for %s: duration input "
-                               "types must be time units in the form of "
-                               "'number:unit'" % e.value)
+            raise RuntimeError(
+                "Invalid duration value for %s: duration input types must "
+                "be time units in the form of 'number:unit'" % e.value)
 
         try:
             if 'gradient' in val:
                 output['gradient'] = {
-                    'top': Unit.fromstring(val['gradient']['top']),
-                    'bottom': Unit.fromstring(val['gradient']['bottom'])
+                    'top': Unit(val['gradient']['top']),
+                    'bottom': Unit(val['gradient']['bottom'])
                 }
             else:
-                output['temperature'] = Unit.fromstring(val['temperature'])
+                output['temperature'] = Unit(val['temperature'])
         except UnitError as e:
-            raise RuntimeError("Invalid temperature value for %s: "
-                               "thermocycle temperature input types must be "
-                               "temperature units in the form of "
-                               "'number:unit'" % e.value)
+            raise RuntimeError(
+                "Invalid temperature value for %s: thermocycle temperature "
+                "input types must be temperature units in the form of "
+                "'number:unit'" % e.value)
 
         if 'read' in val:
             output['read'] = val['read']
@@ -334,7 +334,7 @@ class Manifest(object):
     object : JSON object
         A manifest.json file with the following format:
 
-        .. code-block:: json
+        .. code-block:: none
 
             {
               "format": "python",
@@ -348,8 +348,9 @@ class Manifest(object):
                   "preview": {
                     "refs":{},
                     "parameters": {},
-                  "inputs": {},
-                  "dependencies": []
+                    "inputs": {},
+                    "dependencies": []
+                  }
                 }
               ]
             }
@@ -357,13 +358,9 @@ class Manifest(object):
     """
 
     def __init__(self, json):
-        '''
-        '''
         self.protocols = json['protocols']
 
     def protocol_info(self, name):
-        '''
-        '''
         try:
             return ProtocolInfo(
                 next(p for p in self.protocols if p['name'] == name))
@@ -398,8 +395,8 @@ def run(fn, protocol_name=None, seal_after_run=True):
         help='JSON-formatted protocol configuration file')
     parser.add_argument(
         '--dye_test',
-        help=("Execute protocol by pre-filling preview aliquots with OrangeG "
-              "dye, and provisioning water only."),
+        help='Execute protocol by pre-filling preview aliquots with OrangeG '
+             'dye, and provisioning water only.',
         action="store_true")
     args = parser.parse_args()
 
@@ -409,8 +406,8 @@ def run(fn, protocol_name=None, seal_after_run=True):
         manifest_json = io.open('manifest.json', encoding='utf-8').read()
         manifest = Manifest(json.loads(manifest_json))
         params = manifest.protocol_info(protocol_name).parse(protocol, source)
-        # Add dye to preview aliquots if --dye_test included as an
-        # optional argument
+        # Add dye to preview aliquots if --dye_test included as an optional
+        # argument
         if args.dye_test:
             num_dye_steps = _add_dye_to_preview_refs(protocol)
     else:
@@ -420,8 +417,8 @@ def run(fn, protocol_name=None, seal_after_run=True):
         fn(protocol, params)
         if seal_after_run:
             seal_on_store(protocol)
-        # Convert all provisions to water if --dye_test is included as
-        # an optional argument
+        # Convert all provisions to water if --dye_test is included as an
+        # optional argument
         if args.dye_test:
             _convert_provision_instructions(protocol, num_dye_steps,
                                             len(protocol.instructions) - 1)
@@ -449,12 +446,12 @@ def _add_dye_to_preview_refs(protocol, rs=_DYE_TEST_RS["dye4000"]):
     for ref_name, ref_obj in protocol.refs.items():
 
         ref_cont = ref_obj.container
-        # Raise RuntimeError if any refs have an id, to avoid
-        # adding dye to real samples
+        # Raise RuntimeError if any refs have an id, to avoid adding dye to
+        # real samples
         if ref_cont.id:
-            raise RuntimeError("Cannot run a dye test when any ref has a "
-                               "defined container id. Please resubmit using "
-                               "only new containers.")
+            raise RuntimeError(
+                "Cannot run a dye test when any ref has a defined container "
+                "id. Please resubmit using only new containers.")
 
         # Add dye to each well
         for well in ref_cont.all_wells():
@@ -467,63 +464,69 @@ def _add_dye_to_preview_refs(protocol, rs=_DYE_TEST_RS["dye4000"]):
     return len(protocol.instructions) - starting_num
 
 
-def _convert_provision_instructions(protocol, first_index,
-                                    last_index, rs=_DYE_TEST_RS["water"]):
+def _convert_provision_instructions(protocol, first_index, last_index,
+                                    rs=_DYE_TEST_RS["water"]):
     # Make sure inputs are valid
     if not isinstance(first_index, int):
         raise ValueError("first_index must be a non-negative integer")
     if not isinstance(last_index, int):
         raise ValueError("last_index must be a non-negative integer")
     if first_index < 0:
-        raise ValueError("Indices out of range. First_index must be 0 "
-                         "or greater")
+        raise ValueError(
+            "Indices out of range. first_index must be 0 or greater")
     if first_index > len(protocol.instructions) - 1:
-        raise ValueError("Indices out of range. The last instruction "
-                         "index in the protocol is %d"
-                         "" % (len(protocol.instructions) - 1))
+        raise ValueError(
+            "Indices out of range. The last instruction index in the protocol "
+            "is %d" % (len(protocol.instructions) - 1))
     if last_index > len(protocol.instructions) - 1:
-        raise ValueError("Indices out of range. The last instruction "
-                         "index in the protocol is %d"
-                         "" % (len(protocol.instructions) - 1))
+        raise ValueError(
+            "Indices out of range. The last instruction index in the protocol "
+            "is %d" % (len(protocol.instructions) - 1))
     if last_index < first_index:
-        raise ValueError("last_index must be greater than or equal to "
-                         "first_index")
+        raise ValueError(
+            "last_index must be greater than or equal to first_index")
+
     for instruction in protocol.instructions[first_index:last_index + 1]:
         if instruction.op == "provision":
             instruction.data["resource_id"] = rs
 
 
-def _convert_dispense_instructions(protocol, first_index,
-                                   last_index, rs=_DYE_TEST_RS["water"]):
+def _convert_dispense_instructions(protocol, first_index, last_index,
+                                   rs=_DYE_TEST_RS["water"]):
     # Make sure inputs are valid
     if not isinstance(first_index, int):
         raise ValueError("first_index must be a non-negative integer")
     if not isinstance(last_index, int):
         raise ValueError("last_index must be a non-negative integer")
     if first_index < 0:
-        raise ValueError("Indices out of range. first_index must be 0 "
-                         "or greater")
+        raise ValueError(
+            "Indices out of range. first_index must be 0 or greater")
     if first_index > len(protocol.instructions) - 1:
-        raise ValueError("Indices out of range. The last instruction index "
-                         "in the protocol is %d"
-                         "" % (len(protocol.instructions) - 1))
+        raise ValueError(
+            "Indices out of range. The last instruction index in the protocol "
+            "is %d" % (len(protocol.instructions) - 1))
     if last_index > len(protocol.instructions) - 1:
-        raise ValueError("Indices out of range. The last instruction index "
-                         "in the protocol is %d"
-                         "" % (len(protocol.instructions) - 1))
+        raise ValueError(
+            "Indices out of range. The last instruction index in the protocol "
+            "is %d" % (len(protocol.instructions) - 1))
     if last_index < first_index:
-        raise ValueError("last_index must be greater than or equal "
-                         "to first_index")
+        raise ValueError(
+            "last_index must be greater than or equal to first_index")
+
     for instruction in protocol.instructions[first_index:last_index + 1]:
         if instruction.op == "dispense":
-            instruction.data.pop("reagent", None)
-            instruction.data["resource_id"] = rs
+            if "resource_id" in instruction.data:
+                instruction.data["resource_id"] = rs
+            if "reagent" in instruction.data:
+                instruction.data.pop("reagent", None)
+                instruction.data["resource_id"] = rs
 
 
 def _thermocycle_error_text():
     """
     Returns formatted error text for thermocycle value errors
     """
+
     return """Thermocycle input types must take a list of dictionaries in the form of:
   [{"cycles": integer,
     "steps": [{